The type of a literal floating point number should be double, not float

+++ lib/piglet/field/literal.rb
@@ -26,7 +26,7 @@ module Piglet
         when Integer
           :int
         when Numeric
-          :float
+          :double
         else
           :bytearray
         end
